rearranged the functions. Moved to operations.hpp
